user authentication added

+++ src/resolvers/mutation.js
@@ -1,11 +1,15 @@
 const bcrypt = require('bcrypt');
 const jwt = require('jsonwebtoken');
-const { AuthenticatinError, ForbiddenError } = require('apollo-server-express');
+const { AuthenticationError, ForbiddenError } = require('apollo-server-express');
 const mongoose = require('mongoose');
 require('dotenv').config()
 
 module.exports = {
-    addHouse: async (parent, { title, description = "null", price, location, houseType, images = "" }, { models }) => {
+    addHouse: async (parent, { title, description = "null", price, location, houseType, images = "" }, { models, user }) => {
+
+        if (!user) {
+            throw new AuthenticationError('You must be signed in to create a note');
+        }
 
         if (!title || !price || !location || !houseType) {
             throw new Error("Missing required fields. Ensure all required inputs are provided.");
@@ -19,15 +23,26 @@ module.exports = {
             location,
             houseType,
             images,
-            owner: "APS",//change to user later
+            owner: new mongoose.Types.ObjectId(user.id)
         });
 
 
     },
 
-    deleteHouse: async (parent, { houseId }, { models }) => {
+    deleteHouse: async (parent, { houseId }, { models,user }) => {
+
+        if (!user) {
+            throw new AuthenticationError('You must be signed in to delete a note');
+        }
+        console.log("done1")
+        const house = await models.House.findById(houseId);
+
+        console.log("done2")
+        if(house && String(house.owner)!==user.id){
+            throw new ForbiddenError("You don't have permissions to delete the note");
+        }
+
         try {
-            const house = await models.House.findById(houseId);
             await house.deleteOne();
             console.log("House removed successfully");
             return true;
